Feat : 이미지 리사이징 준비
현재 프론트에서 리사이징을 해서 주는 상태인데 더 필요시 쉽게 수정할 수 있게 구조 변경
