+++ backend/src/app.module.ts
@@ -1,6 +1,6 @@
 import { MiddlewareConsumer, Module, NestModule } from "@nestjs/common";
 import { TypeOrmModule } from "@nestjs/typeorm";
-import { typeORMConfig } from "./configs/typeorm.config";
+import { typeORMConfig } from "./configs";
 import { UserModule } from "./user/user.module";
 import { AuthModule } from "./auth/auth.module";
 import { GroupModule } from "./group/group.module";
