+++ backend/src/configs/index.ts
@@ -1,0 +1,2 @@
+export { typeORMConfig } from "./typeorm.config";
+export { multerOption } from "./s3.config";
