+++ backend/src/custom/myFile/customFile.ts
@@ -1,3 +1,8 @@
 export interface CustomFile extends Express.Multer.File {
   location: string;
+  transforms: Location[];
+}
+
+class Location {
+  location: string;
 }
