+++ backend/src/group/controller/group.controller.ts
@@ -12,7 +12,7 @@ import { CreateGroupResponseDto } from "src/dto/group/createGroupResponse.dto";
 import { UpdateGroupInfoResponseDto } from "src/dto/group/updateGroupInfoResponse.dto";
 import { GetHashTagsResponseDto } from "src/dto/group/getHashTagsResponse.dto";
 import { FileInterceptor } from "@nestjs/platform-express";
-import { multerOption } from "src/image/service/image.service";
+import { multerOption } from "src/configs";
 import {
   SwaggerCreateGroup,
   SwaggerAttendGroup,
